fix normalization and testing
